oss_framework/requirements.txt
# Core Data Stack
duckdb>=0.9.0
dbt-duckdb>=1.6.0
dlt[all]>=0.5.0

# Data Processing
pandas>=2.0.0
pyarrow>=12.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Database Connectors
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Web Framework (for API endpoints)
flask>=2.3.0
fastapi>=0.100.0

# Orchestration
dagster>=1.5.0
dagster-dbt>=0.21.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
tqdm>=4.65.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Security
cryptography>=41.0.0
passlib>=1.7.0
